Reorganize HF tests
